requirements-dev.txt
# Development dependencies for EKS Cluster Validator
# This file combines production and test dependencies for development

# Production dependencies
boto3>=1.34.0
kubernetes>=29.0.0
click>=8.1.0
jinja2>=3.1.0
pyyaml>=6.0.0
requests>=2.31.0
python-dotenv>=1.0.0
colorama>=0.4.6
tabulate>=0.9.0
rich>=13.7.0
pydantic>=2.5.0
loguru>=0.7.0

# Test dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-xdist>=3.0.0
responses>=0.23.0
freezegun>=1.2.0
faker>=15.0.0
coverage>=7.0.0
coverage-badge>=1.1.0

# Code quality and linting
mypy>=1.0.0
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
bandit>=1.7.0
safety>=2.3.0

# Documentation
yamllint>=1.32.0

# Development tools
pre-commit>=3.0.0
pip-tools>=7.0.0

# Build tools
build>=1.0.0
twine>=4.0.0

# Security scanning
semgrep>=1.0.0
